Remove some redundant GL validation.
ANGLE itself can manage most of WebGL validation. This makes the
transition to WebGL 2 validation easier because we won't need to write
new logic for all the new entry points.

To this effect this enables robust memory features in ANGLE, and also
updates the error set logic.

This contribution is funded by http://higharc.com/